Split code to have one component per file; address linter warnings; update snapshots
