+++ __test__/CalloutAlert.test.js
@@ -3,7 +3,6 @@ import Enzyme from 'enzyme'
 import renderer from 'react-test-renderer'
 import { shallow } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import fetchMock from 'fetch-mock'
 
 import CalloutAlert from '../src/CalloutAlert'
