+++ __test__/FetchLoader.test.js
@@ -1,8 +1,9 @@
 import React from 'react'
-import renderer from 'react-test-renderer'
 import Enzyme from 'enzyme'
 import { shallow } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
+
+import '@babel/polyfill'
 import fetchMock from 'fetch-mock'
 
 import { getRandomInt } from './TestUtils'
